% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi.cdf.plot.R
\name{bi.cdf.plot}
\alias{bi.cdf.plot}
\title{Bivariate CDF estimation - plot}
\usage{
bi.cdf.plot(RV, grid_size)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables}

\item{grid_size}{Scalar : The size of the grid on which the plot will be made.}
}
\value{
Returns the CDF estimate across the grid.
}
\description{
Bivariate CDF estimation - plot
}
\examples{
OBJ = bi.cdf.plot(RV = cbind(rnorm(1000),rnorm(1000)), grid_size = 100)
OBJ$Plot3D
OBJ$Contour
}
