% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi.cdf.R
\name{bi.cdf}
\alias{bi.cdf}
\title{Bivariate CDF estimation}
\usage{
bi.cdf(RV, q)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}

\item{q}{Vector: Quantiles.}
}
\value{
Returns the CDF estimate for the quantiles of interest.
}
\description{
Bivariate CDF estimation
}
\examples{
bi.cdf(RV = cbind(rnorm(1000),rnorm(1000)), q = c(0,0))
}
