% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi.pdf.plot.R
\name{bi.pdf.plot}
\alias{bi.pdf.plot}
\title{Bivariate PDF estimation - plot}
\usage{
bi.pdf.plot(RV, grid_size)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}

\item{grid_size}{Scalar : The size of the grid on which the plot will be made.}
}
\value{
Returns the PDF estimate across the grid.
}
\description{
Bivariate PDF estimation - plot
}
\examples{
OBJ = bi.pdf.plot(RV = cbind(rnorm(1000),rnorm(1000)), grid_size = 100)
OBJ$Plot3D
OBJ$Contour
}
