% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi.pdf.R
\name{bi.pdf}
\alias{bi.pdf}
\title{Bivariate PDF estimation}
\usage{
bi.pdf(RV, q)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables}

\item{q}{Vector: Quantiles}
}
\value{
Returns the PDF estimate for the quantiles of interest.
}
\description{
Bivariate PDF estimation
}
\examples{
bi.pdf(RV = cbind(rnorm(1000),rnorm(1000)), q = c(0,0))
}
