% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bitcoin.Ethereum.R
\docType{data}
\name{Bitcoin.Ethereum}
\alias{Bitcoin.Ethereum}
\title{Bitcoin and Ethereum returns 2017-08-19 / 2020-04-23}
\format{
An object of class \code{data.frame} with 979 rows and 3 columns.
}
\usage{
data(Bitcoin.Ethereum)
}
\description{
Bitcoin and Ethereum returns 2017-08-19 / 2020-04-23
}
\references{
https://www.binance.com/en
}
\keyword{datasets}
