% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blomqvists.beta.R
\name{blomqvists.beta}
\alias{blomqvists.beta}
\title{Blomqvist's beta based on copula CDF estimation}
\usage{
blomqvists.beta(RV)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}
}
\value{
Returns the Blomqvist's beta based on copula CDF estimation.
}
\description{
Blomqvist's beta based on copula CDF estimation
}
\examples{
blomqvists.beta(RV = cbind(rnorm(1000),rnorm(1000)))
}
