% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.cdf.plot.R
\name{copula.cdf.plot}
\alias{copula.cdf.plot}
\title{Bivariate CDF estimation of copula - plot}
\usage{
copula.cdf.plot(RV, grid_size)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}

\item{grid_size}{Scalar : The size of the grid on which the plot will be made.}
}
\value{
Returns the CDF estimate across the grid.
}
\description{
Bivariate CDF estimation of copula - plot
}
\examples{
OBJ = copula.cdf.plot(RV = cbind(rnorm(1000),rnorm(1000)), grid_size = 100)
OBJ$Plot3D
OBJ$Contour
}
