% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.cdf.R
\name{copula.cdf}
\alias{copula.cdf}
\title{Bivariate CDF estimation of copula}
\usage{
copula.cdf(RV, q)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}

\item{q}{Vector: Quantiles belonging to (0,1)}
}
\value{
Returns the CDF estimate for the quantiles of interest.
}
\description{
Bivariate CDF estimation of copula
}
\examples{
copula.cdf(RV = cbind(rnorm(1000),rnorm(1000)), q = c(0.5,0.5))
}
