% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.pdf.plot.R
\name{copula.pdf.plot}
\alias{copula.pdf.plot}
\title{Bivariate PDF estimation of copula - plot}
\usage{
copula.pdf.plot(RV, grid_size, sqrt = FALSE)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}

\item{grid_size}{Scalar : The size of the grid on which the plot will be made.}

\item{sqrt}{if sqrt = TRUE copula estimate is shown on the square root scale. It helps to have a better look at some of the details in the PDF. However, the estimate is misleading.}
}
\value{
Returns the PDF estimate across the grid.
}
\description{
Bivariate PDF estimation of copula - plot
}
\examples{
OBJ = copula.pdf.plot(RV = cbind(rnorm(1000),rnorm(1000)), grid_size = 100)
OBJ$Plot3D
OBJ$Contour
}
