% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.pdf.R
\name{copula.pdf}
\alias{copula.pdf}
\title{Bivariate PDF estimation of copula}
\usage{
copula.pdf(RV, q)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}

\item{q}{Vector: Quantiles belonging to ]0,1[}
}
\value{
Returns the PDF estimate for the quantiles of interest.
}
\description{
Bivariate PDF estimation of copula
}
\examples{
copula.pdf(RV = cbind(rnorm(1000),rnorm(1000)), q = c(0.5,0.5))
}
