% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginis.gamma.R
\name{ginis.gamma}
\alias{ginis.gamma}
\title{Gini's gamma based on copula CDF estimation}
\usage{
ginis.gamma(RV)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}
}
\value{
Returns the Gini's gamma based on copula CDF estimation.
}
\description{
Gini's gamma based on copula CDF estimation
}
\examples{
ginis.gamma(RV = cbind(rnorm(1000),rnorm(1000)))
}
