% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCD.R
\name{KCD}
\alias{KCD}
\title{Kernel Copula Dependance}
\usage{
KCD(RV)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}
}
\value{
Returns the Kernel Copula Dependance based on copula CDF estimation.
}
\description{
Kernel Copula Dependance
}
\examples{
KCD(RV = cbind(rnorm(1000),rnorm(1000)))
}
