% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kpobs.R
\name{Kpobs}
\alias{Kpobs}
\title{Kernel Pseudo-Observations}
\usage{
Kpobs(X)
}
\arguments{
\item{X}{n x d-matrix (or d-vector) of random variates to be converted to pseudo-observations.}
}
\value{
Compute the Kernel pseudo-observations for the given data matrix.
}
\description{
Kernel Pseudo-Observations
}
\examples{
Kpobs(X = rnorm(10000))
}
