% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrant.dependence.R
\name{quadrant.dependence}
\alias{quadrant.dependence}
\title{Quadrant dependence based on copula CDF estimation}
\usage{
quadrant.dependence(RV, q)
}
\arguments{
\item{RV}{Matrix: Realizations of two random variables.}

\item{q}{Vector: Quantiles belonging to (0,1)}
}
\value{
Returns the Quadrant dependence based on copula CDF estimation.
}
\description{
Quadrant dependence based on copula CDF estimation
}
\examples{
quadrant.dependence(RV = cbind(rnorm(1000),rnorm(1000)), q = c(0.1,0.1))
}
