% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un.cdf.plot.R
\name{un.cdf.plot}
\alias{un.cdf.plot}
\title{Univariate CDF estimation - plot}
\usage{
un.cdf.plot(RVa, grid_size)
}
\arguments{
\item{RVa}{Vector: Realizations of a random variable}

\item{grid_size}{Scalar : The size of the grid on which the plot will be made.}
}
\value{
Returns the CDF estimate across the grid.
}
\description{
Univariate CDF estimation - plot
}
\examples{
OBJ = un.cdf.plot(RVa = rnorm(10000), grid_size = 1000)
OBJ$Plot
}
