% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un.cdf.R
\name{un.cdf}
\alias{un.cdf}
\title{Univariate CDF estimation}
\usage{
un.cdf(RVa, q)
}
\arguments{
\item{RVa}{Vector: Realizations of a random variable}

\item{q}{Scalar: Quantile}
}
\value{
Returns the CDF estimate for the quantile of interest.
}
\description{
Univariate CDF estimation
}
\examples{
un.cdf(RVa = rnorm(10000), q = 0)
}
