% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un.pdf.plot.R
\name{un.pdf.plot}
\alias{un.pdf.plot}
\title{Univariate PDF estimation - plot}
\usage{
un.pdf.plot(RVa, grid_size)
}
\arguments{
\item{RVa}{Vector: Realizations of a random variable}

\item{grid_size}{Scalar : The size of the grid on which the plot will be made.}
}
\value{
Returns the PDF estimate across the grid.
}
\description{
Univariate PDF estimation - plot
}
\examples{
OBJ = un.pdf.plot(RVa = rnorm(10000), grid_size = 1000)
OBJ$Plot
}
