% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un.pdf.R
\name{un.pdf}
\alias{un.pdf}
\title{Univariate PDF estimation}
\usage{
un.pdf(RVa, q)
}
\arguments{
\item{RVa}{Vector: Realizations of a random variable}

\item{q}{Scalar: Quantile}
}
\value{
Returns the PDF estimate for the quantile of interest.
}
\description{
Univariate PDF estimation
}
\examples{
un.pdf(RVa = rnorm(10000), q = 0)
}
